refactor: move PathQuery::merge onto the versioned sub-file dispatch pattern
The merge slot's behavior was branched inline throughout one function
body (if merge_version >= 1 / < 2 / match merge_version), against the
repo convention that version gates live in <op>/{mod.rs,v0.rs,...}
sub-files with a match dispatcher. PathQuery::merge now keeps only the
version-independent prelude (unknown-slot fail-closed validation,
empty-input rejection, the read-mode refusal, the single-input
shortcut) and dispatches to query/merge/{v0,v1,v2}.rs:

- v0: frozen GROVE_V1..V3 behavior — directions silently dropped, all
  limits and offsets refused;
- v1: direction agreement + propagation, limits still refused (kept as
  its own arm — unlike the folded path_query_push intermediate, the
  0/1/2 numbering is spelled out in the version-table docs);
- v2 (GROVE_V4): v1's direction rules plus the per-instance limit lift
  with its exclusive-graft refusals.

Pure restructure — every version's behavior is byte-for-byte what the
inline branches produced, pinned by the untouched merge test suite.
